Extension update emails gh-116
ExtensionsUpdate task: finish implementing enqueueEmail
